+++ skawa_components/lib/markdown_editor/markdown_editor.dart
@@ -36,6 +36,11 @@ abstract class SkawaEditor {
 /// Base class for an editor that uses a textarea as input
 /// and a Div as a render output
 abstract class TextareaEditorBase implements SkawaEditor {
+  @Input()
+  bool disabled = false;
+
+  StreamController<String> get _modeController;
+
   set renderTarget(EditorRenderTarget newTarget);
 
   EditorRenderTarget get renderTarget;
@@ -56,6 +61,7 @@ abstract class TextareaEditorBase implements SkawaEditor {
       renderTarget.updateRender(value, classes: newClasses);
     }
     _displayMode = mode;
+    _modeController.add(_displayMode);
   }
 }
 
@@ -86,8 +92,10 @@ abstract class TextareaEditorBase implements SkawaEditor {
     selector: 'skawa-markdown-editor',
     templateUrl: 'markdown_editor.html',
     styleUrls: const ['markdown_editor.css'],
-    directives: const [AutoFocusDirective, EditorRenderSource, EditorRenderTarget, LanguageDirectionDirective, NgClass])
-class SkawaMarkdownEditorComponent extends TextareaEditorBase implements OnInit, AfterViewInit {
+    directives: const [AutoFocusDirective, EditorRenderSource, EditorRenderTarget, LanguageDirectionDirective, NgClass],
+    changeDetection: ChangeDetectionStrategy.OnPush)
+class SkawaMarkdownEditorComponent extends TextareaEditorBase implements OnInit, AfterViewInit, OnDestroy {
+  final StreamController<String> _modeController = new StreamController<String>();
   final ViewContainerRef containerRef;
   final ChangeDetectorRef changeDetectorRef;
 
@@ -123,6 +131,9 @@ class SkawaMarkdownEditorComponent extends TextareaEditorBase implements OnInit,
   @Output('update')
   Stream get onUpdated => renderSource.onUpdated;
 
+  @Output('mode')
+  Stream<String> get onMode => _modeController.stream;
+
   @override
   String get value => renderSource.value;
 
@@ -141,7 +152,7 @@ class SkawaMarkdownEditorComponent extends TextareaEditorBase implements OnInit,
   String get mode => displayMode;
 
   void editMarkdown() {
-    displayMode = EditorMode.EDIT;
+    if (!disabled) displayMode = EditorMode.EDIT;
   }
 
   void previewMarkdown() {
@@ -179,6 +190,9 @@ class SkawaMarkdownEditorComponent extends TextareaEditorBase implements OnInit,
     if (value != '') renderTarget.updateRender(value, classes: [_emulatedCssClass]);
     renderTarget.htmlElement;
   }
+
+  @override
+  void ngOnDestroy() => _modeController.close();
 
   void _clonePlaceholderIntoTemplate() {
     renderTarget.htmlElement.children.clear();